[refactor] #5 UserRepository Context 의존성 분리
