+++ app/src/main/java/com/yhj/gdscandroidstudy/di/appModule.kt
@@ -1,6 +1,7 @@
 package com.yhj.gdscandroidstudy.di // ktlint-disable filename
 
 import com.yhj.gdscandroidstudy.data.UserRepository
+import com.yhj.gdscandroidstudy.data.UserRepositoryImpl
 import com.yhj.gdscandroidstudy.ui.edit.EditViewModel
 import com.yhj.gdscandroidstudy.ui.home.HomeViewModel
 import com.yhj.gdscandroidstudy.ui.mypage.MyPageViewModel
@@ -8,7 +9,7 @@ import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val appModule = module {
-    single { UserRepository(get()) }
+    single<UserRepository> { UserRepositoryImpl(get()) }
     viewModel { EditViewModel(get()) }
     viewModel { MyPageViewModel(get()) }
     viewModel { HomeViewModel() }
